Add openstack module
Adds openstack module with functions to manage
- client
- endpoints
- project
- role
- service
- user

E.g. used in KeystoneService to register users for service
components.

Usage is to use NewOpenStack() to get a new instance of OpenStack
with an authenticated client. Then can do o.CreateEndpoint(...)
to create a new service endpoint.

Implemented as a new module to not have dependency to gopher
on remaining lib-common content, where these are not used.