Initial Ceph support
This patch adds the initial shared struct and functions to enable common
Ceph support in OSP operators. Given that most of the operators share
the same Ceph backend data structure and utilities, it's useful to start
this module and avoid a lot of code duplication.

The structure of the module is supposed to have:

- types: where the data structure is defined (and can be eventually
  extended)

- funcs: where the functions are exported and can be used by the
  operators

- util: where utility functions are implemented

Signed-off-by: Francesco Pantano <fpantano@redhat.com>