Add envtest CRD helper
In envtest we need to be able to install all the CRDs the operator
under test depends on.

This patch adds test.GetCRDDirFromModule helper to get the location of
CRDs from the operator go module defining it.

It also adds GetOpenShiftCRDDir function to look up the OpenShift CRDs.
As those are not published we store a copy of them in lib-common.