Move OpenStackServiceBase API struct to lib-common/api module
Each OpenStack service CRD will has a set of common fields like
ContainerImage, Replicas, NodeSelector, etc. So factor these common
fields out to a struct and move it to a common place. It is suggested
that each operator import this struct and embed it to the Spec struct
via:

	api.OpenStackServiceBase `json:",inline"`

The embedding and the inline tags makes this change almost fully backward
compatible with the original direct fields. The CRD will not change as
it will have the fields from OpenStackServiceBase directly inlined.
The controller code that reads Spec fields does not have to change as
the fields are directly accessible in the Spec struct due to unambiguous
embedding. However in the code where the Spec is instantiated (e.g. in
the tests) the Spec instantiation needs to use the

    OpenStackServiceBase: api.OpenStackServiceBase{...}

syntax to populate the embedded struct fields.