Fix hashing of ConfigMaps
Whenever an operator creates a new `ConfigMap` using methods
`createOrPatchConfigMap` or `createOrGetCustomConfigMap` the calculated
has will not be correct.

Due to this incorrect hash we end up seeing 1 unnecessary destroy/create
cycle of the pods.

In the Cinder operator where this was observed we would see that at
least the cinder-volume, cinder-backup, and cinder-scheduler containers
would successfully run their init containers and then the services would
start running displaying their logs, but a couple of seconds later they
would just be terminated just to do the cycle again.

After 1 terminate/create cycle they would stabilize.

We could see that their respective `StatefulSet` had a `generation`
count of 2 (when it should be 1) and this was caused by a change of the
`ConfigMap` hash.

The problem comes from a difference in the hash calculation of a newly
created `ConfigMap` and the same one when retrieved from OpenShift.

This happens because when we create the `ConfigMap` it will locally set
the `TypeMeta` fields to empty strings (even if we set it otherwise),
but when we get the `ConfigMap` from OpenShift it will have `Kind` and
`APIVersion` set.

This patch solves the issue by only hashing the actual data from the
`ConfigMap`, fields `Data` and `BinaryData` instead of the whole
`ConfigMap`.