Fix hashing of Secrets
This applies the same way to fix hashing of ConfigMaps
( commit a00d28b2af98ed239a5371a0012d83abd1654a1a ) to Secrets.

Instead of calc the hash for the full Secret object the hash now gets
only calculated for the Data, StringData and Type of the Secret. This
prevents consuming resources to e.g. restart if the hash changes
because the ObjectMetadata of the Secret changed.