Extra volumes support to lib-common/modules/storage
This patch introduces the required data structures and functions needed
by the OpenStack operators to mount Extra Volumes other than the default
ones required by each service. Extra Volumes can be Secrets, ConfigMaps
and any other kubernetes entity that can be mounted to pods. The main
storage module provides the basic building block for this feature, and
the operator related data structures (moved to the API section for each
operatorsm are extended to include additional parameters that can be
used as conditions in the operator logic. The Propagation function is
able to process both internal (the ability to propagate an extraVolume
within the single operator pods) and external propagation (the ability
to allocate and mount a volume across different storage operators
(Glance, Cinder, Manila, Ansible, etc), which implements a way to filter
the volumes a pod is supposed to mount according to the propagation
policy defined in the main crd. Volumes and VolumeMounts are two tied
structures, and they should be processed the same way when we perform
propagation.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>