[Job]Remove Job deletion from DoJob
We have a logic error around Job deletion.

1. controller calls DoJob() on lib-common's Job
2. lib-common waits for the Job to finish and if succeeds, it deletes the Job.
3. the controller calculates the hash of the Job and puts it to Status.Hash
4. the controller calls Status.Update() to persist the hash to the DB

If step 4. fails due to a conflict then the reconcile is requeued and the
controller will see that no Job exists (as it was deleted at 2.) and
recreates it. This means we can potentially run the same Job multiple
times if we are unlucky.

So this changes removed the automatic job deletion logic from DoJob and
instead provides a new DeleteAllSucceededJobs() call to the controllers
to use to clean up the Jobs.

*This is a breaking change*

The callers should do the following to adapt:
1) After DoJob() returned ctrl.Result{}, nil, (i.e. the Job succeeded)
   the callers should store the hash of the Job in its Status
2) The caller should persist the Status.
3) If persisting the Status succeeded then call DeleteAllSucceededJobs
   providing the hash of the Jobs to be deleted.