Default TTLSecondsAfterFinished on Jobs
Currently lib-common automatically and immediately deletes the Job after
it finished if preserve is set to false. However
* if the caller of DoJob fails to persist the fact that the Job is done
  (e.g. due to conflict)
OR
* the caller's next reconcile reads stale data from cache seeing that the
  Job is not finished but also does not exists as it was already deleted

Then the client has no way to decide that the Job does not need to be
run again. This could lead to duplicate Job execution.

So this patch delays the automatic Job deletion with a default TTL of 10
minutes counted from the finish of the Job. We believe that this is
plenty of time to allow the caller to successfully persist the Job state
even if it has to retry it and to let the operator-sdk update the cache.

See the previous fix attempt[1] (a partial fix) and the nova issue [2] about
the stale read for more details.

[1] https://github.com/openstack-k8s-operators/lib-common/pull/77
[2] https://github.com/openstack-k8s-operators/nova-operator/issues/110

Closes: https://github.com/openstack-k8s-operators/nova-operator/issues/110