Do not include data.Port in the keystone endpoint
Keystone endpoint points to the route created on top of the specific
service, which is exposed on data.Port.
However, having the port in the endpoint url let the openstack cli
to fail because it tries to reach the route on the port hardcoded in
the url.
This change just removes the port from the generated endpointurl.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>