Only try to update TTL on an existing Job
With #78 lib-common uses TTL on Jobs to clean them up and to support
PreserveJob the TTL can be change on a Job even after it is finished.
However #78 has a bug as it tries to CreateOrPatch the Job _after_ it is
deleted but with an empty Job definition. This fails with

  "error": "an empty namespace may not be set during creation"

This patch makes sure that lib-common only tries to update the TTL if
the Job exists.