Add networkattachment pkg
Adds networkattachment pkg with helpers
* GetNADWithName to get an NAD with its name,
* CreateNetworksAnnotation to create the annotation which can be added
to a deployment spec to attach additional interfaces to the depl pods.
This is the replacement for GetNADAnnotation from annotations pkg.
* GetNetworkStatusFromAnnotation which returns the network-status from
a pods annotation,
* VerifyNetworkStatusFromAnnotation verifies if NetworkStatus annotation
for the pods of a deployment, pods identified via the service label
selector, matches the passed in network attachments and the number of
per network IPs the ready count of the deployment. Return true if count
matches with the list of IPs per network.