Mirror order conditions as False, Unknown, True
This is a breaking change on Conditions.Mirror(). Originally this call
ordered conditions as False, True, Unknown then picked the first of that
list to reflect. However all the use cases (we see) assume that if there
is reasourceA that depends on the status of resourceB then resourceA
reconciler calls:

  resourceB.Status.Mirror(ResourceBSpecificSubConditionOfA)

and expects that the call returns the most sever non True condition if any
or return a True condition otherwise. So the original behavior that
returned True even if there was Unknown condition was unexpected.

This patch changes the order of the status to False, Unknown, True. But
keeps the original logic that if the ReadyCondition is True then the
rest of the conditions are ignored even if there are False conditions in
the list.