Do not requeue for running Job
Polling the Job status unnecessary causing k8s API load and noisy logging.
The caller of DoJob() should declare that it owns a Job via
Owns(&batchv1.Job{}) in its SetupWithManager call. Then the caller will
be automatically reconciled when the Job status is changed.

There is an edge case when the Job is just created and immediately read
back from the API causing NotFound temporarily. In this case returning
the NotFound error is simpler than asking for and explicit Requeue and
it will still trigger the automatic requeue.