[pre-commit]skip branch check in CI
We run all pre-commit on both push to branch and open pull request
events. This means that we run it also against master after a PR is
approved. In that case the no-commit-to-branch check, that enforces no
direct commit to master branch fails, so we disable this check in CI.
Locally it still make sense to keep it as it is good git hygiene to
always work on a local topic branch.