Allow customize route via route.Override
Adds an OverridSpec to the Route which allows to customize
metadata.Annotations, metadata.Labels and spec of a route. The
override values get merged into the object definition created by
the operator. This allows e.g. to add custom labels, configure
the route via annotations as in [1], or set TLS parameters.

[1] https://docs.openshift.com/container-platform/4.13/networking/routes/route-configuration.html#nw-route-specific-annotations_route-configuration

Jira: OSP-21715
Jira: OSP-26299