Allow customize service via service.Override
Adds an OverridSpec to the Service which allows to customize
metadata.Annotations, metadata.Labels and spec of a service. The
override values get merged into the object definition created by
the operator. This is part of moving the external connection
config into the openstack-operator and allows to override the
default service created by an operatior with a MetalLB LoadBalancer
service spec.

Jira: OSP-26690