Give specific domain id for user/project
Keystone supports "domain" and users, groups and projects belong to
a specific domain, and it's possible that multiple entries with
the same name exist in different domains.

This introduces the interface to specify the domain where the user or
the project should exist, to avoid unexpected behavior caused by such
"duplicate" entities.