Merge Annotations and Labels in CreateOrPatch
Identified in OCP4.13 where metallb adds an annotation to the
service when handling it, it resulted in reconcile loop since
the current service CreateOrPatch did not expect that annotations
will be altered outside the owner.

This changes the different functions to merge Labels and Annotations
in CreateOr Patch instead of assigning.