[TLS] Don't set commonName and only use subjectAltNames
commonName has a max length of 64 chars, which easy can be reached
in route DNS names scheme.

From [1]
~~~
RFC 2818, published in May 2000, deprecates the use of the Common
Name (CN) field in HTTPS certificates for subject name verification.
Instead, it recommends using the “Subject Alternative Name” extension
(SAN) of the “dns name” type.
~~~

Lets not set the commonName an just use the subjectAltNames for our
certs.

[1] https://cloud.redhat.com/blog/details-on-https-common-name-deprecation-in-openshift-4.10