test: allow fetching CRD from a local replace statement
Ginkgo test suites require fetching the CRDs of the custom resources
used in the tests. Those are fetched from the go.mod and any replace
statement is parsed and honoured.
Add the ability to process replace statements that point to a local
directory as well, to ease the developement process when a project
depends on a feature from another project that has not been merged yet.