lowercase module/crd path
This change modifes how we calculate the path for a module
to allow github usernames with any case to funciton.
the previous code only work for all lowercase usernames

the usernames in github are canonically cases senseitve even
though any case is accpeted in the url.

The path on disk used by go will however encoded to lowercase.
This change makes our code have the same behvior.

We reuse the same path encoding golang uses[1] when downloads the package
to the local cache.

https://github.com/golang/go/blob/c54bc3448390d4ae4495d6d2c03c9dd4111b08f1/src/cmd/go/internal/module/module.go#L421