Add .ObservedGeneration to statefulset and deployments test module
In order to be able to rollout the full '.ObservedGeneration' pattern we need
envTest to pass if this is added as a condition to mark a Deployment or
StatefulSet Ready. This change introduces this parameter for the test module
so we can take this parameter into account in the envTests.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>