Skip encodePath based on runtime GOOS
Qualify of life improvement that skips path encoding on systems
where it's likely to cause test failures - Macs, it impacts Macs.

Look, I appreciate that this is a selfish change and only impacts a small
subset of engineers. But it's annoying to edit this file and get tests
passing locally each time.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>