Fix TestValidateRoutedOverrides
ValidateRoutedOverrides does not guarantee a specific order of the
returned ErrorList. As a result the "Multiple wrong override endpoints"
sometimes fails even the expected errors are in the list. This changes
to use ContainElements as it does not matter which order the errors
are in the returned ErrorlList from ValidateRoutedOverrides.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>
(cherry picked from commit 024880500e1e571978216504ddb31f0a21afdd53)