Create getDNSClusterDomain
Openshift DNS creates it's records using the value in clusterDomain.
This information could be customizable in the futre.

Currently on most of the operators they're using a hardcoded value,
this function allows to all operators get this information from
lib-common, and in case in the future this value is customizable
we'd only need to change it here.