Expose PodAffinity and PodAntiAffinity struct and build overrides
This patch introduces a very basic struct to expose Pod Affinity/Antiaffinity
and NodeAffinity interfaces as part of the Topology CR spec and it allows to
patch the (opinionated) default that is currently applied to the services
through DistributePods function call.
The existing DistributePods function is still present in the code to ensure
backward compatibility and not block lib-common bumps, but is marked as
deprecated.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>