Simplify the Affinity struct exposed to service operators
This patch introduces some more wrapping around corev1.PodAffinitySpec
to simplify what is exposed in the service operators. In particular,
it is possible to consume the AffinityOverrides struct that can
implement either affinity or antiaffinity rules. They uses the same
structs behind the scenes (PodAffinityTerm and WeightedPodAffinityTerm),
for both RequiredDuringSchedulingIgnoredDuringExecution and
PreferredDuringSchedulingIgnoredDuringExecution, but they have a
different semantic when included in the corev1.Affinity k8s object.
The old behavior of distributing Pods is currently preserved for
operators where we do not want to provide this interface.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>