[conditions] Reset conditions on Init()
Recently the operators changed to use CreateServiceReadyCondition
instead of ExposeServiceReadyCondition. An update of an existing
environment would result in the ExposeServiceReadyCondition to be
never removed from the conditions list of a CR because the list
is not reset on Init().

This adds a Reset() method and a call right at the start of Init().

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>