Add TopoRef struct and common functions
TopoRef represents a struct referenced by the service operators to
retrieve a Topology that is eventually applied to the resulting
Deployments/StatefulSets.
This patch allows to reduce the amount of code duplication across the
operators, providing both the TopoRef struct that can be added in the
Services' API, as well as "EnsureTopologyRef" and "EnsureDeletedTopologyRef"
where the basic logic of retrieving and managing finalizers is implemented.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>