Add GetLabelSelector utility function
This function is used as utility to get the list of "matchLabels" used
as selector for Deployments and StatefulSet. Most of the service
operators use to pass a "serviceLabels map[string]string", which is
available in the controller(s). This function translates the data
structure into a metav1.LabelSelector that we can easily plug to the
existing ensureTopology logic provided by infra-operator [1].

[1] https://github.com/openstack-k8s-operators/infra-operator/blob/main/apis/topology/v1beta1/topology_types.go#L126

Signed-off-by: Francesco Pantano <fpantano@redhat.com>