Do not use MustParse with wrong StorageRequest size
In case of a wrong user input, the operator crashes without any chance
to recover.
The MustParse implementation wraps the ParseQuantity function, that
instead of panic() returns an error that can be easily handled at the
operator level. This change does not alter the current behavior, but
it adds a preliminary check that does not let the validation to continue
if the quantity is not valid. It also moves away from MustParse in favor
of ParseQuantity. By doing this we can get a webhook feedback about the
wrong quantity instead of checking the operators log and see the panic
stacktrace.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>