[Deployment] Add PollRolloutStatus and poll deployment on updates
When a deployment gets updated (config, image or changed spec) and
the rollout results in the deployment to fail it should be reflected
and the caller be able to evaluate. Currently if there is an issue
like the above, the deployment rollout/update results in DeadlineExceed
but it is not reflected in the service operators. Since the old
config replicas are still up and healthy, the service is still
functional with the old specs, but the rollout issue is not reflected.

This adds PollRolloutStatus() and calls it from CreateOrPatch()
when the oparation != create.

Using rolloutPollInterval and rolloutPollTimeout, the caller can
control the interval and timeout of one poll run which uses
PollUntilContextTimeout().

The status and message of each run is reflected as d.rolloutStatus
and d.rolloutMessage. Status can be Complete, Progressing and
ProgressDeadlineExceeded. The message has a corresponding string to
each of the status.

To note, the DeadlineExceeded from PollUntilContextTimeout() in
CreateOrPatch() is ignored to allow the caller to reflect the
rolleoutStatus/rolloutMessage in its DeploymentCondition and
reconcile and trigger a new poll.

Jira: OSPRH-14472

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>