[envTest] Add AssertVolumeMountPathExists function
We recently introduced extraMounts envTests in several operators, and
they rely on the common helpers/volume Assert function to check if the
resulting deployment has the right Volume/VolumeMountPath.
This patch introduces a new version of AssertVolumeMountExists to either
assert a subPath or a regular MountPath (or both!) exists.

Jira: https://issues.redhat.com/browse/OSPRH-15839

Signed-off-by: Francesco Pantano <fpantano@redhat.com>